+++ sources/packages/api/src/route-controllers/institution/models/institution.dto.ts
@@ -1,7 +1,6 @@
 import { IsNotEmpty, IsOptional } from "class-validator";
 import { PartialType } from "@nestjs/mapped-types";
 import { BCeIDDetailsDto } from "../../../route-controllers/user/models/bceid-account.dto";
-import { Institution } from "../../../database/entities";
 
 /**
  * DTO object for institution creation.
